fetched data and displayed it
tested that data was fetched from teh external source and that it was being displayed on chrome.

expected output: name, discription, price, and photo displayed on page. this is what I saw

to test that It was getting the data from the external source I changed foodToBuy.item to foodToBuy.itee to make sure that undfined was displayed and it was