button
added the addbutton top the rest of the foods and tested on chrome, one from each section. so pancakes from breakfast, the salad from dinner andthe popsciles from dessert.

the expected and correct output was to have each food that was added to teh menu, visible in the menu section.

this was successfull