appropriate messages
expected output: when there are no foods in menu "No food in menu" will be displayed and when there are foods in menu "there are * foods in your menu" will be displayed telling the user how many foods they have in their menu.
this is the correct output.

this is what is visible to the user . Yesss. success!!