remove food from cart
expected output was that when the remove button is pressed that specifc food from the cart is removed.

I suck

that is not hwat happened. when a food is added to teh cart, it creates 2 remove buttons and only one works and it only removes every  food in the cart expect the first one.

idk how to fix that