yesss
yess
yess
 yess
 yess
 yess
 yes

hahahha
 yes

expected output= when the remove button is clicked for its respective food that food should remove itself from cart.

this is what happened when tested on chrome, so the remove function and button is now fixed and the user is able to remove which ever item from their cart that they like.

eg:

teh user cliks the 2nd food remove button and the second food is removed from the cart

invaild testing-
tested on chrome that when I added 5 foods to cart, and clicked the remove first food button, only teh first food was removed which is the expected output (thank god), this means that I fixed the previous problem of removing everything up to the first food when the first foods remove buttion is clicked.