invalid idek boundary
expected output: when food is added to menu the add to menu button for that speficc food gets disabled. this makes aboundary so that only one of each food can be added to menu.

I tested this on chrome by adding foods from all sections (brekkie, dins, and dessert). at first only the dinner menu did the expected output, this was because I had not put the disable button code in the other sections. So after I put the disable button code in all sections the expected oiut put worken when I addn the pancakes and salad and popciles, only these foods-add to menu button was disablled.