boundary/limit esting
when there are no foods in the menu the user should not have the ablility to remove nothing from the menu. there should not be the remove food from menu button available to the user unless there is a food in the menu.

I tested the on chrome and had successful results. when there were 0 foods in the menu, the user had no option to remove anything from the menu, b ut as soon as a food is added to the menu the remove button appears for that specifc food.