added a card
texpected and boundary tested and de buggeged the program

added card to contain the food info and picture to simplify the code on the main page.

tested addinng a food to mthe menu( the food output is correct as the card works and allows the user to see all the food info and pictures) and then that food is added to the menu. when trying to add that food again the button should disable. it does